Added Player to groups
Added the Player class to groups "updatables" and
"drawables", and used those in the game loop for
easier management of game objects for later.

Also fixed an issue with accidentally calling
clock.tick() twice, so I also updated the
PLAYER_SPEED constant, because the game runs
twice as fast now.