Asteroid destruction added
Now the bullets actually destroy the asteroids. Added another collision
check in the game loop for every bullet on screen to see if any of them
collide with an asteroid.

Also had to update the variable inside the shoot() method of Player to
an instance variable so I could use it as a comparison in main.py.