Hotfix #1
I realised why the asteroid circles kept being entirely filled out.
Now this is fixed, by updating the asteroid's draw method, which uses
a builtin pygame method to draw a circle.

I realised this method also can take in an optional width parameter,
which specifies the line width to draw the circles with.

This was also said in the course but I somehow missed it. Oh well, I
guess better late than never.